requirements.txt
datasets==3.6.0
ipdb==0.13.13
torch==2.7.0
transformers==4.52.4
accelerate==1.7.0
peft==0.15.2
rich==14.0.0
vllm==0.9.1
trl==0.18.1
openai==1.86.0
litellm==1.72.4
pyzmq==27.0.0
tqdm==4.67.1
numpy==2.2.6
scipy==1.15.3
matplotlib==3.10.3
flashinfer-python==0.2.2